+++ wxfile.cpp
@@ -485,9 +485,9 @@ bool MainFrame::ExtractZipEntry(const wxString& zippath,
             // read and write in chunks so we can show a progress dialog
             const int BUFFER_SIZE = 4000;
             char buf[BUFFER_SIZE];
-            int incount = 0;
-            int outcount = 0;
-            int lastread, lastwrite;
+            size_t incount = 0;
+            size_t outcount = 0;
+            size_t lastread, lastwrite;
             double filesize = (double) entry->GetSize();
             if (filesize <= 0.0) filesize = -1.0;        // show indeterminate progress
             
@@ -505,7 +505,7 @@ bool MainFrame::ExtractZipEntry(const wxString& zippath,
                   break;
                }
                char msg[128];
-               sprintf(msg, "File size: %.2g MB", double(incount) / 1048576.0);
+               sprintf(msg, "File size: %.2f MB", double(incount) / 1048576.0);
                if (AbortProgress((double)incount / filesize, wxString(msg,wxConvLocal))) {
                   outcount = 0;
                   break;
